requirements.txt
Flask==0.10.1
Jinja2==2.6
Werkzeug==0.8.3
gunicorn==19.3.0
Flask-SQLAlchemy==2.1
Flask-Migrate==1.6.0
Flask-Script==2.0.5
psycopg2==2.6.1
slackweb==1.0.5